+++ sync-for-payroll/src/sdk/models/operations/createjournal.ts
@@ -3,13 +3,10 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
 
 export class CreateJournalRequest extends SpeakeasyBase {
-    @SpeakeasyMetadata({ data: "request, media_type=application/json" })
-    journal?: shared.Journal;
-
     /**
      * Allow a sync upon push completion.
      */
@@ -27,6 +24,9 @@ export class CreateJournalRequest extends SpeakeasyBase {
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
+
+    @SpeakeasyMetadata({ data: "request, media_type=application/json" })
+    journalPrototype?: shared.JournalPrototype;
 
     /**
      * Time limit for the push operation to complete before it is timed out.